Move some code out of the waymark::messages into its own crate
