+++ crates/bin/fuzzer/Cargo.toml
@@ -14,6 +14,7 @@ uuid = { workspace = true, features = ["serde", "v4"] }
 tokio = { workspace = true }
 waymark = { workspace = true }
 waymark-dag = { workspace = true }
+waymark-proto = { workspace = true }
 waymark-ir-parser = { workspace = true }
 waymark-runner-state = { workspace = true }
 waymark-backend-memory = { workspace = true }
