+++ crates/bin/soak-harness/Cargo.toml
@@ -11,7 +11,7 @@ waymark-runner-state = { workspace = true }
 waymark-backend-postgres = { workspace = true }
 waymark-backend-postgres-migrations = { workspace = true }
 waymark-workflow-registry-backend = { workspace = true }
-waymark = { workspace = true }
+waymark-proto = { workspace = true }
 
 anyhow = { workspace = true }
 chrono = { workspace = true }
@@ -24,5 +24,5 @@ uuid = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 sqlx = { workspace = true }
-tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
+tokio = { workspace = true, features = ["macros", "rt-multi-thread", "process", "signal"] }
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
