+++ crates/lib/message-conversions/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "waymark-message-conversions"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+serde_json = { workspace = true }
+waymark-proto = { workspace = true }
